+++ tests/helper.py
@@ -3,7 +3,8 @@ import tensorflow as tf
 __all__ = ['assert_variables']
 
 
-def assert_variables(names, exist=True, trainable=None, scope=None):
+def assert_variables(names, exist=True, trainable=None, scope=None,
+                     collections=None):
     """
     Assert variables of `name_or_names` meet certain criterion.
 
@@ -12,6 +13,8 @@ def assert_variables(names, exist=True, trainable=None, scope=None):
         exist (bool): Assert variables exist or not.
         trainable: Assert variables are trainable or not.
         scope (None or str): The scope prefix to be prepended to the names.
+        collections (Iterable[str]): Additional graph collections, where
+            to ensure the variables are in.
     """
     def normalize_name(n):
         return n.rsplit(':', 1)[0]
@@ -26,6 +29,11 @@ def assert_variables(names, exist=True, trainable=None, scope=None):
                    for v in tf.global_variables()}
     trainable_vars = {normalize_name(v.name): v
                       for v in tf.trainable_variables()}
+    collections = list(collections or ())
+    collection_vars = [
+        {normalize_name(v.name): v for v in tf.get_collection(c)}
+        for c in collections
+    ]
 
     for name in names:
         if exist:
@@ -44,6 +52,13 @@ def assert_variables(names, exist=True, trainable=None, scope=None):
                     raise AssertionError('Variable `{}` is expected to be '
                                          'trainable, but turned out not to be '
                                          'trainable'.format(name))
+
+            # check collections
+            for coll, coll_vars in zip(collections, collection_vars):
+                if name not in coll_vars:
+                    raise AssertionError('Variable `{}` is expected to be '
+                                         'in the collection `{}`, but turned '
+                                         'out not.'.format(name, coll))
 
         else:
             if name in global_vars: