+++ tfsnippet/utils/__init__.py
@@ -10,6 +10,7 @@ from .doc_utils import *
 from .imported import *
 from .invertible_matrix import *
 from .misc import *
+from .model_vars import *
 from .random import *
 from .reuse import *
 from .scope import *
@@ -37,20 +38,20 @@ __all__ = [
     'ensure_variables_initialized', 'flatten_to_ndims', 'get_batch_size',
     'get_cache_root', 'get_config_defaults', 'get_config_validator',
     'get_default_scope_name', 'get_default_session_or_error',
-    'get_dimensions_size', 'get_rank', 'get_reuse_stack_top', 'get_shape',
-    'get_static_shape', 'get_uninitialized_variables', 'get_variable_ddi',
-    'get_variables_as_dict', 'global_reuse', 'humanize_duration',
-    'instance_reuse', 'is_assertion_enabled', 'is_float', 'is_integer',
-    'is_shape_equal', 'is_tensor_object',
-    'is_tensorflow_version_higher_or_equal', 'iter_files', 'makedirs',
-    'maybe_check_numerics', 'maybe_close', 'minibatch_slices_iterator',
-    'register_config_arguments', 'register_config_validator',
-    'register_tensor_wrapper_class', 'reopen_variable_scope', 'reshape_tail',
-    'resolve_negative_axis', 'root_variable_scope', 'scoped_set_config',
-    'set_assertion_enabled', 'set_cache_root', 'set_check_numerics',
-    'set_random_seed', 'settings', 'should_check_numerics',
-    'split_numpy_array', 'split_numpy_arrays', 'transpose_conv2d_axis',
-    'transpose_conv2d_channels_last_to_x',
+    'get_dimensions_size', 'get_model_variables', 'get_rank',
+    'get_reuse_stack_top', 'get_shape', 'get_static_shape',
+    'get_uninitialized_variables', 'get_variable_ddi', 'get_variables_as_dict',
+    'global_reuse', 'humanize_duration', 'instance_reuse',
+    'is_assertion_enabled', 'is_float', 'is_integer', 'is_shape_equal',
+    'is_tensor_object', 'is_tensorflow_version_higher_or_equal', 'iter_files',
+    'makedirs', 'maybe_check_numerics', 'maybe_close',
+    'minibatch_slices_iterator', 'model_variable', 'register_config_arguments',
+    'register_config_validator', 'register_tensor_wrapper_class',
+    'reopen_variable_scope', 'reshape_tail', 'resolve_negative_axis',
+    'root_variable_scope', 'scoped_set_config', 'set_assertion_enabled',
+    'set_cache_root', 'set_check_numerics', 'set_random_seed', 'settings',
+    'should_check_numerics', 'split_numpy_array', 'split_numpy_arrays',
+    'transpose_conv2d_axis', 'transpose_conv2d_channels_last_to_x',
     'transpose_conv2d_channels_x_to_last', 'unflatten_from_ndims',
     'validate_enum_arg', 'validate_group_ndims_arg', 'validate_int_tuple_arg',
     'validate_n_samples_arg', 'validate_positive_int_arg',